requirements.txt
flask
gunicorn
pypdf2
google-generativeai
python-dotenv